Doctor view implemented
