+++ src/main.tsx
@@ -3,6 +3,7 @@ import { createRoot } from "react-dom/client";
 import { HeroUIProvider } from "@heroui/react";
 import "./index.css";
 import App from "./App.tsx";
+import "@modules/appointment-management/utils/icon.ts";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
