+++ src/modules/appointment-management/components/doctor/CurrentAttention/NoShift/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./NoShift";
