+++ src/modules/appointment-management/components/doctor/CurrentAttention/ShiftInfo/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ShiftInfo";
