+++ src/modules/appointment-management/components/doctor/CurrentAttention/index.tsx
@@ -1,2 +1,3 @@
 export { default } from "./CurrentAttention";
-export { default as InfoItem } from "./InfoItem";
+export { default as ShiftInfo } from "./ShiftInfo";
+export { default as NoShift } from "./NoShift";
