+++ src/modules/appointment-management/utils/icon.ts
@@ -4,6 +4,17 @@ import {
   faUser,
   faCalendar,
 } from "@fortawesome/free-regular-svg-icons";
-import { faStethoscope, faArrowLeft } from "@fortawesome/free-solid-svg-icons";
+import {
+  faStethoscope,
+  faArrowLeft,
+  faUserCheck,
+} from "@fortawesome/free-solid-svg-icons";
 
-library.add(faCircleCheck, faUser, faStethoscope, faCalendar, faArrowLeft);
+library.add(
+  faCircleCheck,
+  faUser,
+  faStethoscope,
+  faCalendar,
+  faArrowLeft,
+  faUserCheck
+);
